requirements.txt
Pillow
pyocr
flask